rearanging the models more logically between packages
